feat(cli): `wait`/`-w` flag displays fancy spinner
- cli: The `--wait` flag is available to `stop,run,start` commands. It
  now displays a nice spinner group displaying state for each container
  provessed individually.
- cni: perform cni detach on container stop/kill events while netns is
  still present. Otherwise cni doesn not cleanup uptable chains
  correctly resulting in leaked cni networks on the host

+++ cmd/bbctl/stop/container.go
@@ -3,8 +3,11 @@ package stop
 import (
 	"context"
 	"fmt"
+	"strings"
+	"time"
 
 	"github.com/amimof/blipblop/pkg/client"
+	"github.com/amimof/blipblop/pkg/cmdutil"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
@@ -43,19 +46,82 @@ func NewCmdStopContainer(cfg *client.Config) *cobra.Command {
 				}
 			}()
 
-			// Send stop request to server
-			for _, cname := range args {
+			// Send stop or kill for each container in args without waiting
+			if !viper.GetBool("wait") {
+				for _, cname := range args {
 
-				if viper.GetBool("force") {
-					if _, err = c.ContainerV1().Kill(ctx, cname); err != nil {
-						logrus.Fatal(err)
+					if viper.GetBool("force") {
+						if _, err = c.ContainerV1().Kill(ctx, cname); err != nil {
+							logrus.Fatal(err)
+						}
+					} else {
+						if _, err = c.ContainerV1().Stop(ctx, cname); err != nil {
+							logrus.Fatal(err)
+						}
 					}
+
+					fmt.Printf("requested to stop container %s\n", cname)
 				}
-				if _, err = c.ContainerV1().Stop(ctx, cname); err != nil {
-					logrus.Fatal(err)
+			}
+
+			// Send stop or kill for each container in args and wait for them all to stop
+			if viper.GetBool("wait") {
+				dash := cmdutil.NewDashboard(args)
+				go dash.Loop(ctx)
+
+				for i, cname := range args {
+					// Fire off start operations concurrently
+					go func(idx int, containerID string) {
+						if viper.GetBool("force") {
+							if _, err = c.ContainerV1().Kill(ctx, cname); err != nil {
+								dash.FailMsg(idx, fmt.Sprintf("failed to start: %v", err))
+								return
+							}
+						} else {
+							if _, err = c.ContainerV1().Stop(ctx, cname); err != nil {
+								dash.FailMsg(idx, fmt.Sprintf("failed to start: %v", err))
+								return
+							}
+						}
+
+						dash.Update(idx, func(s *cmdutil.ServiceState) {
+							s.Text = "stopping…"
+						})
+
+						// Continously check container
+						for {
+
+							dash.FailAfterMsg(idx, viper.GetDuration("timeout"), "failed to start in time")
+
+							ctr, werr := c.ContainerV1().Get(ctx, containerID)
+							if werr != nil {
+								dash.FailMsg(idx, werr.Error())
+								return
+							}
+
+							phase := ctr.GetStatus().GetPhase().GetValue()
+							dash.Update(idx, func(s *cmdutil.ServiceState) {
+								s.Text = fmt.Sprintf("%s…", phase)
+							})
+
+							if phase == "stopped" {
+								dash.DoneMsg(idx, "stopped successfully")
+								return
+							}
+
+							if strings.Contains(phase, "Err") {
+								dash.FailMsg(idx, "failed to start")
+								return
+							}
+
+							// Wait until retry
+							time.Sleep(250 * time.Millisecond)
+						}
+					}(i, cname)
 				}
 
-				fmt.Printf("requested to stop container %s\n", cname)
+				dash.WaitAnd(cancel)
+
 			}
 		},
 	}
